Commit 4 sorts
подправил сцены, произвел сортировку внутри сцен. изменил пару скриптов (произвел наследование от Splashes к TeleportWithEffect). доработал дождь.
