--- Assets/Material/New Material.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c693207811850c40a19a0842cbf6b5e
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
