--- Assets/Material/Sun2.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b038fb99551ab0f4eab46e203ccc1793
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
