--- Assets/Material/Sun3.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7ce05acba3ba2e34c848e63858f11a95
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
