--- Assets/Material/lvl1platform.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 59eee2f8cbbf3ee468cbd2d032e11e47
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
