--- Assets/Material/lvl2platform.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d1ca326b5620bbf4e9250bfe35536b29
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
