+++ Assets/Material/rain.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 64a21c09c54ace4489b6aa2b2ce1f8d0
-TextScriptImporter:
+guid: da9236314f43b694ea0f0438d5b8232f
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
