--- Assets/Plagins/Joystick Pack/Documentaion.pdf.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: cfa813ed406827e43b967abd54ab5b9a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 107631
-  packageName: Joystick Pack
-  packageVersion: 2.1
-  assetPath: Assets/Joystick Pack/Documentaion.pdf
-  uploadId: 301683
