--- Assets/Plagins/Joystick Pack/Examples.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5a71259a40885244a8fda03175513ba5
-folderAsset: yes
-timeCreated: 1513537895
-licenseType: Store
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
