--- Assets/Plagins/Joystick Pack/Examples/Example Scene.unity.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 414bddd31c5ea4f4c9f5d1452e5bbb41
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 107631
-  packageName: Joystick Pack
-  packageVersion: 2.1
-  assetPath: Assets/Joystick Pack/Examples/Example Scene.unity
-  uploadId: 301683
