--- Assets/Plagins/Joystick Pack/Examples/JoystickPlayerExample.cs
@@ -1,8 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class JoystickPlayerExample : MonoBehaviour
-{
-     
-}
