--- Assets/Plagins/Joystick Pack/Examples/JoystickPlayerExample.cs.meta
@@ -1,18 +1,0 @@
-fileFormatVersion: 2
-guid: a37d705c8e9115e439ec799186182634
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 107631
-  packageName: Joystick Pack
-  packageVersion: 2.1
-  assetPath: Assets/Joystick Pack/Examples/JoystickPlayerExample.cs
-  uploadId: 301683
