--- Assets/Plagins/Joystick Pack/Examples/JoystickSetterExample.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class JoystickSetterExample : MonoBehaviour
-{
-    
-}
