--- Assets/Plagins/Joystick Pack/Examples/JoystickSetterExample.cs.meta
@@ -1,18 +1,0 @@
-fileFormatVersion: 2
-guid: 3d73bf03f88f371469fd7b43fc97d77e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 107631
-  packageName: Joystick Pack
-  packageVersion: 2.1
-  assetPath: Assets/Joystick Pack/Examples/JoystickSetterExample.cs
-  uploadId: 301683
