--- Assets/Plagins/Joystick Pack/Examples/Player.mat.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-guid: 59e4faf24927edf4a82f463ad16f1c37
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-AssetOrigin:
-  serializedVersion: 1
-  productId: 107631
-  packageName: Joystick Pack
-  packageVersion: 2.1
-  assetPath: Assets/Joystick Pack/Examples/Player.mat
-  uploadId: 301683
