+++ Assets/Prefab/minirain.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e51d30842e1e8a6459ed1ab9ea2f5071
-DefaultImporter:
+guid: 8df32c1b8e777dc4a9234649229a0a97
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
