+++ Assets/Prefab/rain.prefab
@@ -13,8 +13,9 @@ GameObject:
   - component: {fileID: 2923587115018226241}
   - component: {fileID: 3384445618257933907}
   - component: {fileID: -5253571585568311345}
+  - component: {fileID: -3635887734984590354}
   m_Layer: 0
-  m_Name: drop
+  m_Name: rain
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -65,7 +66,7 @@ MeshRenderer:
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
-  - {fileID: 2100000, guid: 31321ba15b8f8eb4c954353edc038b1d, type: 2}
+  - {fileID: 2100000, guid: da9236314f43b694ea0f0438d5b8232f, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -135,3 +136,19 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &-3635887734984590354
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6314530503876186578}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5db5c4b5e92c851488fc23896f56715f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  particlePrefab: {fileID: 6314530503876186578, guid: 8df32c1b8e777dc4a9234649229a0a97, type: 3}
+  explosionForce: 5
+  timeDestroy: 0.85
+  particleCount: 5
