+++ Assets/Scenes/lvl 2.unity.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 6bbb490a51567364a9f675196fac3014
-NativeFormatImporter:
+guid: 829ea13a64dafbb48a0093ab4edd4141
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
