--- Assets/Scripts/Scene_scripts/CameraPivot.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-using UnityEngine.UI; // Подключаем пространство имен UI
-
-public class CameraPivot : MonoBehaviour
-{
-   
-}
