--- Assets/Scripts/Scene_scripts/CameraPivot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ccc7f84234c4d7041917df6cf3d387a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
