+++ Assets/Scripts/Scene_scripts/Finish.cs
@@ -3,10 +3,10 @@ using UnityEngine.SceneManagement;
 
 public class Finish : MonoBehaviour
 {
-
-
+    [SerializeField]
+    private int indexScene;
     private void OnTriggerStay(Collider other)
     {
-        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
+        SceneManager.LoadScene(indexScene);
     }
 }
